+++ frontend/src/components/browser/card/order-by-caption.vue
@@ -43,6 +43,9 @@ export default {
       } else if (DATE_SORT_BY.has(this.orderBy)) {
         const date = new Date(ov);
         ov = DATE_FORMAT.format(date);
+      } else if (this.orderBy == "search_score") {
+        // Round Whoosh float into a two digit integer.
+        ov = Math.round(Number.parseFloat(ov) * 10);
       } else if (TIME_SORT_BY.has(this.orderBy)) {
         const date = new Date(ov);
         // this is what needs v-html to work